// i3-retrieval-monitoring/src/model/purchaseItem.ts
export class PurchaseItem {
  public Field1;
  public Field2;
  public Field3;
  public Field4;
  public Field5;
  public Field6;
  public Field7;
  public Field8;
  public V1;
  public V2;
  public V3;
  public V4;
  public V5;
  public V6;
  public V7;
  public V15;
  public Field16;
  public Field17;
  public Field18;
  public Field19;
  public Field20;
  public Field21;
  public Field23;
  public Field24;
  public rowId;
  public last;

  constructor(fields?: {
    Field1?: string;
    Field2?: string;
    Field3?: string;
    Field4?: string;
    Field5?: string;
    Field6?: string;
    Field7?: string;
    Field8?: string;
    V1?: string;
    V2?: string;
    V3?: string;
    V4?: string;
    V5?: string;
    V6?: string;
    V7?: string;
    V15?: string;
    Field16?: string;
    Field17?: string;
    Field18?: string;
    Field19?: string;
    Field20?: string;
    Field21?: string;
    Field23?: string;
    Field24?: string;
    rowId?: number;
    last?: number;
  }) {
    Object.assign(this, fields);
  }
}